# refac: improve lexing inline elements step's performance