--- a/src/Lexer.ts
+++ b/src/Lexer.ts
@@ -89,8 +89,8 @@ export class _Lexer {
 
     this.blockTokens(src, this.tokens);
 
-    let next;
-    while (next = this.inlineQueue.shift()) {
+    for (let i = 0; i < this.inlineQueue.length; i++) {
+      const next = this.inlineQueue[i];
       this.inlineTokens(next.src, next.tokens);
     }