@UziTech 2023-12-29
The only issue I can see is that `this.inlineQueue` will be an empty array at the end of the while loop and not at the end of the for loop. This could cause issues with some advanced uses of marked