@SpencerWhitehead7 2023-12-29
As far as I could tell, nothing reads the array after the end of the while loop. I don't see Lexer using it anywhere else, or it being accessed elsewhere, and the lexer just returns the array of tokens at the end.

If there's cases I'm missing, it could be reset to an empty array after the for loop runs, or reversed and then popped instead of shifted.