@UziTech 2023-12-29
I don't think it will break anything in marked if it isn't reset but if someone uses the lexer with `new marked.Lexer()` than it could cause issues when they reuse the same lexer.

I think we can just reset it to an empty array after the for loop