@SpencerWhitehead7 2023-12-29
Oh, that makes sense. Thank you for explaining the case so clearly. I added reseting it to an empty array.